+++ sdk/containerregistry/azure-containerregistry/azure/containerregistry/_container_repository_client.py
@@ -5,7 +5,10 @@
 # ------------------------------------
 from typing import TYPE_CHECKING
 
+from azure.core.tracing.decorator import distributed_trace
+
 from ._base_client import ContainerRegistryBaseClient
+from ._helpers import _is_tag
 from ._models import RepositoryProperties, TagProperties, RegistryArtifactProperties
 
 if TYPE_CHECKING:
@@ -29,12 +32,18 @@ class ContainerRepositoryClient(ContainerRegistryBaseClient):
         :returns: None
         :raises: None
         """
-        if not endpoint.startswith("https://"):
+        if not endpoint.startswith("https://") and not endpoint.startswith("http://"):
             endpoint = "https://" + endpoint
         self._endpoint = endpoint
         self.repository = repository
         super(ContainerRepositoryClient, self).__init__(endpoint=self._endpoint, credential=credential, **kwargs)
 
+    def _get_digest_from_tag(self, tag):
+        # type: (str) -> str
+        tag_props = self.get_tag_properties(tag)
+        return tag_props.digest
+
+    @distributed_trace
     def delete(self, **kwargs):
         # type: (...) -> None
         """Delete a repository
@@ -44,36 +53,32 @@ class ContainerRepositoryClient(ContainerRegistryBaseClient):
         """
         self._client.container_registry.delete_repository(self.repository, **kwargs)
 
-    def delete_registry_artifact(self, digest):
+    @distributed_trace
+    def delete_registry_artifact(self, digest, **kwargs):
         # type: (str) -> None
-        """Delete a registry artifact
+        """Delete a registry artifact. A registry artifact can only be deleted from the digest
 
         :param digest: The digest of the artifact to be deleted
         :type digest: str
         :returns: None
         :raises: :class:~azure.core.exceptions.ResourceNotFoundError
         """
-        raise NotImplementedError("Has not been implemented")
+        self._client.container_registry_repository.delete_manifest(self.repository, digest, **kwargs)
 
-    def delete_tag(self, tag):
+    @distributed_trace
+    def delete_tag(self, tag, **kwargs):
         # type: (str) -> None
-        """Delete a tag
+        """Delete a tag from a repository
 
         :param tag: The digest of the artifact to be deleted
         :type tag: str
         :returns: None
         :raises: :class:~azure.core.exceptions.ResourceNotFoundError
         """
-        raise NotImplementedError("Has not been implemented")
+        self._client.container_registry_repository.delete_tag(self.repository, tag, **kwargs)
 
-    def get_digest_from_tag(self, tag):
-        # type: (str) -> str
-        for t in self.list_tags():
-            if t.tag == tag:
-                return t.digest
-        return ""
-
-    def get_properties(self):
+    @distributed_trace
+    def get_properties(self, **kwargs):
         # type: (...) -> RepositoryProperties
         """Get the properties of a repository
 
@@ -81,9 +86,11 @@ class ContainerRepositoryClient(ContainerRegistryBaseClient):
         :raises: None
         """
         # GET '/acr/v1/{name}'
-        resp = self._client.container_registry_repository.get_properties(self.repository)
-        return RepositoryProperties._from_generated(resp)  # pylint: disable=protected-access
+        return RepositoryProperties._from_generated(  # pylint: disable=protected-access
+            self._client.container_registry_repository.get_properties(self.repository, **kwargs)
+        )
 
+    @distributed_trace
     def get_registry_artifact_properties(self, tag_or_digest, **kwargs):
         # type: (str, Dict[str, Any]) -> RegistryArtifactProperties
         """Get the properties of a registry artifact
@@ -93,18 +100,16 @@ class ContainerRepositoryClient(ContainerRegistryBaseClient):
         :returns: :class:~azure.containerregistry.RegistryArtifactProperties
         :raises: :class:~azure.core.exceptions.ResourceNotFoundError
         """
-        # GET '/acr/v1/{name}/_manifests/{digest}'
-        # TODO: If `tag_or_digest` is a tag, need to do a get_tags to find the appropriate digest,
-        # generated code only takes a digest
-        if self._is_tag(tag_or_digest):
-            tag_or_digest = self.get_digest_from_tag(tag_or_digest)
-        # TODO: The returned object from the generated code is not being deserialized properly
+        if _is_tag(tag_or_digest):
+            tag_or_digest = self._get_digest_from_tag(tag_or_digest)
+
         return RegistryArtifactProperties._from_generated(  # pylint: disable=protected-access
             self._client.container_registry_repository.get_registry_artifact_properties(
                 self.repository, tag_or_digest, **kwargs
             )
         )
 
+    @distributed_trace
     def get_tag_properties(self, tag, **kwargs):
         # type: (str, Dict[str, Any]) -> TagProperties
         """Get the properties for a tag
@@ -114,36 +119,39 @@ class ContainerRepositoryClient(ContainerRegistryBaseClient):
         :returns: :class:~azure.containerregistry.TagProperties
         :raises: :class:~azure.core.exceptions.ResourceNotFoundError
         """
-        # GET '/acr/v1/{name}/_tags/{reference}'
         return TagProperties._from_generated(  # pylint: disable=protected-access
             self._client.container_registry_repository.get_tag_properties(self.repository, tag, **kwargs)
         )
 
+    @distributed_trace
     def list_registry_artifacts(self, **kwargs):
         # type: (...) -> ItemPaged[RegistryArtifactProperties]
-        """List the registry artifacts for a repository
+        """List the artifacts for a repository
 
         :keyword last: Query parameter for the last item in the previous query
         :type last: str
-        :keyword n: Max number of items to be returned
-        :type n: int
+        :keyword page_size: Number of items per page
+        :type page_size: int
         :keyword orderby: Order by query parameter
         :type orderby: :class:~azure.containerregistry.RegistryArtifactOrderBy
         :returns: ~azure.core.paging.ItemPaged[RegistryArtifactProperties]
         :raises: None
         """
-        raise NotImplementedError("Not implemented")
-        # TODO: turn this into an ItemPaged
-        # artifacts = self._client.manifests.get_list(
-        #     self.repository,
-        #     last=kwargs.get("last", None),
-        #     n=kwargs.get("n", None),
-        #     orderby=kwargs.get("orderby"),
-        # )  # ,
-        # # cls=lambda objs: [RegistryArtifacts._from_generated(x) for x in objs])
+        last = kwargs.pop("last", None)
+        n = kwargs.pop("page_size", None)
+        orderby = kwargs.pop("order_by", None)
+        return self._client.container_registry_repository.get_manifests(
+            self.repository,
+            last=last,
+            n=n,
+            orderby=orderby,
+            cls=lambda objs: [
+                RegistryArtifactProperties._from_generated(x) for x in objs  # pylint: disable=protected-access
+            ],
+            **kwargs
+        )
 
-        # return RegistryArtifactProperties._from_generated(artifacts)
-
+    @distributed_trace
     def list_tags(self, **kwargs):
         # type: (...) -> ItemPaged[TagProperties]
         """List the tags for a repository
@@ -158,35 +166,42 @@ class ContainerRepositoryClient(ContainerRegistryBaseClient):
         return self._client.container_registry_repository.get_tags(
             self.repository,
             last=kwargs.pop("last", None),
-            n=kwargs.pop("top", None),
+            n=kwargs.pop("page_size", None),
             orderby=kwargs.pop("order_by", None),
             digest=kwargs.pop("digest", None),
             cls=lambda objs: [TagProperties._from_generated(o) for o in objs],  # pylint: disable=protected-access
             **kwargs
         )
 
-    def set_manifest_properties(self, digest, value):
+    @distributed_trace
+    def set_manifest_properties(self, digest, permissions, **kwargs):
         # type: (str, ContentPermissions) -> None
         """Set the properties for a manifest
 
         :param digest: Digest of a manifest
         :type digest: str
-        :param value: The property's values to be set
-        :type value: ContentPermissions
-        :returns: ~azure.core.paging.ItemPaged[TagProperties]
+        :param permissions: The property's values to be set
+        :type permissions: ContentPermissions
+        :returns: None
         :raises: None
         """
-        raise NotImplementedError("Has not been implemented")
 
-    def set_tag_properties(self, tag, permissions):
+        self._client.container_registry_repository.update_manifest_attributes(
+            self.repository, digest, value=permissions._to_generated(), **kwargs  # pylint: disable=protected-access
+        )
+
+    @distributed_trace
+    def set_tag_properties(self, tag, permissions, **kwargs):
         # type: (str, ContentPermissions) -> None
         """Set the properties for a tag
 
         :param tag: Tag to set properties for
         :type tag: str
-        :param value: The property's values to be set
-        :type value: ContentPermissions
-        :returns: ~azure.core.paging.ItemPaged[TagProperties]
+        :param permissions: The property's values to be set
+        :type permissions: ContentPermissions
+        :returns: None
         :raises: None
         """
-        raise NotImplementedError("Has not been implemented")
+        self._client.container_registry_repository.update_tag_attributes(
+            self.repository, tag, value=permissions._to_generated(), **kwargs  # pylint: disable=protected-access
+        )