+++ sdk/identity/azure-identity/azure/identity/aio/_credentials/certificate.py
@@ -7,17 +7,17 @@ from typing import TYPE_CHECKING
 import msal
 
 from .._internal import AadClient, AsyncContextManager
-from .._internal.decorators import log_get_token_async
+from .._internal.get_token_mixin import GetTokenMixin
 from ..._credentials.certificate import get_client_credential
 from ..._internal import AadClientCertificate, validate_tenant_id
-from ..._persistent_cache import _load_persistent_cache, TokenCachePersistenceOptions
+from ..._persistent_cache import _load_persistent_cache
 
 if TYPE_CHECKING:
     from typing import Any, Optional
     from azure.core.credentials import AccessToken
 
 
-class CertificateCredential(AsyncContextManager):
+class CertificateCredential(AsyncContextManager, GetTokenMixin):
     """Authenticates as a service principal using a certificate.
 
     The certificate must have an RSA private key, because this credential signs assertions using RS256.
@@ -57,6 +57,7 @@ class CertificateCredential(AsyncContextManager):
 
         self._client = AadClient(tenant_id, client_id, cache=cache, **kwargs)
         self._client_id = client_id
+        super().__init__()
 
     async def __aenter__(self):
         await self._client.__aenter__()
@@ -67,27 +68,8 @@ class CertificateCredential(AsyncContextManager):
 
         await self._client.__aexit__()
 
-    @log_get_token_async
-    async def get_token(self, *scopes: str, **kwargs: "Any") -> "AccessToken":  # pylint:disable=unused-argument
-        """Asynchronously request an access token for `scopes`.
+    async def _acquire_token_silently(self, *scopes: str) -> "Optional[AccessToken]":
+        return self._client.get_cached_access_token(scopes, query={"client_id": self._client_id})
 
-        This method is called automatically by Azure SDK clients.
-
-        :param str scopes: desired scopes for the access token. This method requires at least one scope.
-        :rtype: :class:`azure.core.credentials.AccessToken`
-        :raises ~azure.core.exceptions.ClientAuthenticationError: authentication failed. The error's ``message``
-          attribute gives a reason. Any error response from Azure Active Directory is available as the error's
-          ``response`` attribute.
-        """
-        if not scopes:
-            raise ValueError("'get_token' requires at least one scope")
-
-        token = self._client.get_cached_access_token(scopes, query={"client_id": self._client_id})
-        if not token:
-            token = await self._client.obtain_token_by_client_certificate(scopes, self._certificate, **kwargs)
-        elif self._client.should_refresh(token):
-            try:
-                await self._client.obtain_token_by_client_certificate(scopes, self._certificate, **kwargs)
-            except Exception:  # pylint: disable=broad-except
-                pass
-        return token
+    async def _request_token(self, *scopes: str, **kwargs: "Any") -> "AccessToken":
+        return await self._client.obtain_token_by_client_certificate(scopes, self._certificate, **kwargs)