+++ sdk/core/azure-core/CHANGELOG.md
@@ -1,14 +1,20 @@
 # Release History
 
-## 1.13.0 (Unreleased)
+## 1.13.0 (2021-04-02)
 
-### Features
+Azure core requires Python 2.7 or Python 3.6+ since this release.
 
+### New Features
+
+- Added `azure.core.utils.parse_connection_string` function to parse connection strings across SDKs, with common validation and support for case insensitive keys.
 - Supported adding custom policies  #16519
+- Added `~azure.core.tracing.Link` that should be used while passing `Links` to  `AbstractSpan`.
+- `AbstractSpan` constructor can now take in additional keyword only args.
 
 ### Bug fixes
 
 - Make NetworkTraceLoggingPolicy show the auth token in plain text. #14191
+- Fixed RetryPolicy overriding default connection timeout with an extreme value #17481
 
 ## 1.12.0 (2021-03-08)